Post discussion changes
